RSE-347: Create civicase-base module
