+++ ang/civicase-base.js
@@ -1,0 +1,3 @@
+(function (angular, $, _) {
+  angular.module('civicase-base', CRM.angRequires('civicase-base'));
+})(angular, CRM.$, CRM._);
