+++ ang/civicase-base/services/activity-category.service.js
@@ -1,0 +1,14 @@
+(function (angular, $, _, CRM) {
+  var module = angular.module('civicase-base');
+
+  module.service('ActivityCategory', ActivityCategory);
+
+  /**
+   * Activity Category Service
+   */
+  function ActivityCategory () {
+    this.getAll = function () {
+      return CRM['civicase-base'].activityCategories;
+    };
+  }
+})(angular, CRM.$, CRM._, CRM);
