+++ ang/civicase-base/services/case-type.service.js
@@ -1,0 +1,14 @@
+(function (angular, $, _, CRM) {
+  var module = angular.module('civicase-base');
+
+  module.service('CaseType', CaseType);
+
+  /**
+   * CaseType Service
+   */
+  function CaseType () {
+    this.getAll = function () {
+      return CRM['civicase-base'].caseTypes;
+    };
+  }
+})(angular, CRM.$, CRM._, CRM);
