+++ ang/civicase-base/services/custom-activity-field.service.js
@@ -1,0 +1,14 @@
+(function (angular, $, _, CRM) {
+  var module = angular.module('civicase-base');
+
+  module.service('CustomActivityField', CustomActivityField);
+
+  /**
+   * Tagset Service
+   */
+  function CustomActivityField () {
+    this.getAll = function () {
+      return CRM['civicase-base'].customActivityFields;
+    };
+  }
+})(angular, CRM.$, CRM._, CRM);
