+++ ang/civicase-base/services/file-category.service.js
@@ -1,0 +1,14 @@
+(function (angular, $, _, CRM) {
+  var module = angular.module('civicase-base');
+
+  module.service('FileCategory', FileCategory);
+
+  /**
+   * Activity Types Service
+   */
+  function FileCategory () {
+    this.getAll = function () {
+      return CRM['civicase-base'].fileCategories;
+    };
+  }
+})(angular, CRM.$, CRM._, CRM);
