+++ ang/civicase-base/services/priority.service.js
@@ -1,0 +1,14 @@
+(function (angular, $, _, CRM) {
+  var module = angular.module('civicase-base');
+
+  module.service('Priority', Priority);
+
+  /**
+   * Priority Service
+   */
+  function Priority () {
+    this.getAll = function () {
+      return CRM['civicase-base'].priority;
+    };
+  }
+})(angular, CRM.$, CRM._, CRM);
